refactor(genui): export standard Node handler
