+++ .github/genui-server.instructions.md
@@ -10,4 +10,4 @@ Keep shared agent-service contracts and helpers in `service/common`. `ChatMessag
 
 Use `app/common/sse.ts` for standard SSE frames and response headers. Pass event IDs or additional headers through its options instead of cloning the SSE framing and header literals in individual routes.
 
-Build `genui-server` as an ESM Node.js service through `rslib.config.ts`. Keep protocol handlers based on standard Web `Request` and `Response`, register them in `src/routes.ts`, and keep Node HTTP stream conversion in `src/node-server.ts`. Do not introduce framework-specific request or response types into protocol routes.
+Build `genui-server` as an ESM Node.js service through `rslib.config.ts`. Export the deployment entry as the standard Node `handler(IncomingMessage, ServerResponse)` function from `src/index.ts`; keep listening-socket ownership in `src/server.ts`. Keep protocol handlers based on standard Web `Request` and `Response`, register them in `src/routes.ts`, and keep Node HTTP stream conversion in `src/node-server.ts`. Do not introduce framework-specific request or response types into protocol routes.
