+++ .gitignore
@@ -67,6 +67,5 @@ target/
 
 # next
 .next
-.vercel
 
 .gradle
