+++ packages/genui/a2ui-prompt/AGENTS.md
@@ -8,9 +8,9 @@ This package may contain thin package-specific adapters such as
 `readA2UICatalogFromDirectory`, but shared prompt rendering and built-in catalog
 behavior should live in the server agent source and be re-exported from here.
 
-The server package must remain self-contained for package-root Vercel
-deployments. Do not make `packages/genui/server` depend on `@lynx-js/genui/a2ui-prompt`
-at runtime.
+The server package must remain self-contained for package-root deployments. Do
+not make `packages/genui/server` depend on `@lynx-js/genui/a2ui-prompt` at
+runtime.
 
 When editing exported functions or constants reachable through this package,
 remember that `isolatedDeclarations` requires explicit types for exported APIs.
