+++ packages/genui/playground/src/mock/openui-scenarios.ts
@@ -1014,7 +1014,7 @@ statsCard = Card([CardHeader("📊  Stats", "At a glance"), statsRow], "sunk", "
 statsRow = Stack([Stack([TextContent("8.4k", "large-heavy"), TextContent("👥  followers", "small")], "column", false, "xs"), Stack([TextContent(142, "large-heavy"), TextContent("📝  posts", "small")], "column", false, "xs"), Stack([TextContent(312, "large-heavy"), TextContent("➡️  following", "small")], "column", false, "xs")], "row", true, "l", "stretch", "start")
 
 bioCard = Card([CardHeader("🪪  About", "A few words"), bioText], "card", "column", false, "m", "start", "start")
-bioText = TextContent("Building thoughtful interfaces with React and TypeScript. Previously at Linear and Vercel. Open-source contributor focused on design systems and developer tooling.", "default")
+bioText = TextContent("Building thoughtful interfaces with React and TypeScript. Previously at Linear and Stripe. Open-source contributor focused on design systems and developer tooling.", "default")
 
 activityCard = Card([CardHeader("🕒  Recent Activity", "What she's been up to"), activityList], "card", "column", false, "m", "stretch", "start")
 activityList = Stack([act1, Separator(), act2, Separator(), act3], "column", false, "s")
@@ -1024,7 +1024,7 @@ act3 = Stack([TextContent("✍️  Published 'Edge rendering with Lynx'", "small
 
 actionRow = Stack([Buttons([Button("➕  Follow", Action([@ToAssistant("Follow Maya Chen")]), "primary", "normal", "large"), Button("✉️  Message", Action([@ToAssistant("Send Maya a message")]), "secondary", "normal", "medium")])], "row", false, "m", "start", "start")`,
     parsed:
-      '{\n  "root": {\n    "type": "element",\n    "typeName": "Stack",\n    "props": {\n      "children": [\n        {\n          "type": "element",\n          "typeName": "Card",\n          "props": {\n            "children": [\n              {\n                "type": "element",\n                "typeName": "CardHeader",\n                "props": {\n                  "title": "👩‍💻  Maya Chen",\n                  "subtitle": "Senior Frontend Engineer"\n                },\n                "partial": false,\n                "hasDynamicProps": false\n              },\n              {\n                "type": "element",\n                "typeName": "Stack",\n                "props": {\n                  "children": [\n                    {\n                      "type": "element",\n                      "typeName": "Tag",\n                      "props": {\n                        "text": "📍  San Francisco"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Tag",\n                      "props": {\n                        "text": "⚛️  React"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Tag",\n                      "props": {\n                        "text": "🧩  Design Systems"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Tag",\n                      "props": {\n                        "text": "⏳  8 yrs exp"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    }\n                  ],\n                  "direction": "row",\n                  "wrap": true,\n                  "gap": "m",\n                  "align": "stretch",\n                  "justify": "start"\n                },\n                "partial": false,\n                "hasDynamicProps": false,\n                "statementId": "profileMeta"\n              }\n            ],\n            "variant": "card",\n            "direction": "column",\n            "wrap": false,\n            "gap": "m",\n            "align": "start",\n            "justify": "start"\n          },\n          "partial": false,\n          "hasDynamicProps": false,\n          "statementId": "profileCard"\n        },\n        {\n          "type": "element",\n          "typeName": "Card",\n          "props": {\n            "children": [\n              {\n                "type": "element",\n                "typeName": "CardHeader",\n                "props": {\n                  "title": "📊  Stats",\n                  "subtitle": "At a glance"\n                },\n                "partial": false,\n                "hasDynamicProps": false\n              },\n              {\n                "type": "element",\n                "typeName": "Stack",\n                "props": {\n                  "children": [\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "8.4k",\n                              "size": "large-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "👥  followers",\n                              "size": "small"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": 142,\n                              "size": "large-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "📝  posts",\n                              "size": "small"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": 312,\n                              "size": "large-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "➡️  following",\n                              "size": "small"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    }\n                  ],\n                  "direction": "row",\n                  "wrap": true,\n                  "gap": "l",\n                  "align": "stretch",\n                  "justify": "start"\n                },\n                "partial": false,\n                "hasDynamicProps": false,\n                "statementId": "statsRow"\n              }\n            ],\n            "variant": "sunk",\n            "direction": "column",\n            "wrap": false,\n            "gap": "m",\n            "align": "start",\n            "justify": "start"\n          },\n          "partial": false,\n          "hasDynamicProps": false,\n          "statementId": "statsCard"\n        },\n        {\n          "type": "element",\n          "typeName": "Card",\n          "props": {\n            "children": [\n              {\n                "type": "element",\n                "typeName": "CardHeader",\n                "props": {\n                  "title": "🪪  About",\n                  "subtitle": "A few words"\n                },\n                "partial": false,\n                "hasDynamicProps": false\n              },\n              {\n                "type": "element",\n                "typeName": "TextContent",\n                "props": {\n                  "text": "Building thoughtful interfaces with React and TypeScript. Previously at Linear and Vercel. Open-source contributor focused on design systems and developer tooling.",\n                  "size": "default"\n                },\n                "partial": false,\n                "hasDynamicProps": false,\n                "statementId": "bioText"\n              }\n            ],\n            "variant": "card",\n            "direction": "column",\n            "wrap": false,\n            "gap": "m",\n            "align": "start",\n            "justify": "start"\n          },\n          "partial": false,\n          "hasDynamicProps": false,\n          "statementId": "bioCard"\n        },\n        {\n          "type": "element",\n          "typeName": "Card",\n          "props": {\n            "children": [\n              {\n                "type": "element",\n                "typeName": "CardHeader",\n                "props": {\n                  "title": "🕒  Recent Activity",\n                  "subtitle": "What she\'s been up to"\n                },\n                "partial": false,\n                "hasDynamicProps": false\n              },\n              {\n                "type": "element",\n                "typeName": "Stack",\n                "props": {\n                  "children": [\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "💬  Replied to a thread on design systems",\n                              "size": "small-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "2 hours ago",\n                              "size": "small"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false,\n                      "statementId": "act1"\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Separator",\n                      "props": {},\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "🚀  Shipped v2.0 of use-open-ui",\n                              "size": "small-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "Yesterday",\n                              "size": "default"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false,\n                      "statementId": "act2"\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Separator",\n                      "props": {},\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "✍️  Published \'Edge rendering with Lynx\'",\n                              "size": "small-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "3 days ago",\n                              "size": "small"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false,\n                      "statementId": "act3"\n                    }\n                  ],\n                  "direction": "column",\n                  "wrap": false,\n                  "gap": "s"\n                },\n                "partial": false,\n                "hasDynamicProps": false,\n                "statementId": "activityList"\n              }\n            ],\n            "variant": "card",\n            "direction": "column",\n            "wrap": false,\n            "gap": "m",\n            "align": "stretch",\n            "justify": "start"\n          },\n          "partial": false,\n          "hasDynamicProps": false,\n          "statementId": "activityCard"\n        },\n        {\n          "type": "element",\n          "typeName": "Stack",\n          "props": {\n            "children": [\n              {\n                "type": "element",\n                "typeName": "Buttons",\n                "props": {\n                  "buttons": [\n                    {\n                      "type": "element",\n                      "typeName": "Button",\n                      "props": {\n                        "label": "➕  Follow",\n                        "action": {\n                          "k": "Comp",\n                          "name": "Action",\n                          "args": [\n                            {\n                              "k": "Arr",\n                              "els": [\n                                {\n                                  "k": "Comp",\n                                  "name": "ToAssistant",\n                                  "args": [\n                                    {\n                                      "k": "Str",\n                                      "v": "Follow Maya Chen"\n                                    }\n                                  ]\n                                }\n                              ]\n                            }\n                          ]\n                        },\n                        "variant": "primary",\n                        "type": "normal",\n                        "size": "large"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": true\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Button",\n                      "props": {\n                        "label": "✉️  Message",\n                        "action": {\n                          "k": "Comp",\n                          "name": "Action",\n                          "args": [\n                            {\n                              "k": "Arr",\n                              "els": [\n                                {\n                                  "k": "Comp",\n                                  "name": "ToAssistant",\n                                  "args": [\n                                    {\n                                      "k": "Str",\n                                      "v": "Send Maya a message"\n                                    }\n                                  ]\n                                }\n                              ]\n                            }\n                          ]\n                        },\n                        "variant": "secondary",\n                        "type": "normal",\n                        "size": "medium"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": true\n                    }\n                  ]\n                },\n                "partial": false,\n                "hasDynamicProps": true\n              }\n            ],\n            "direction": "row",\n            "wrap": false,\n            "gap": "m",\n            "align": "start",\n            "justify": "start"\n          },\n          "partial": false,\n          "hasDynamicProps": true,\n          "statementId": "actionRow"\n        }\n      ],\n      "direction": "column",\n      "wrap": false,\n      "gap": "l"\n    },\n    "partial": false,\n    "hasDynamicProps": true,\n    "statementId": "root"\n  },\n  "meta": {\n    "incomplete": false,\n    "unresolved": [],\n    "orphaned": [],\n    "statementCount": 13,\n    "errors": []\n  },\n  "stateDeclarations": {},\n  "queryStatements": [],\n  "mutationStatements": []\n}',
+      '{\n  "root": {\n    "type": "element",\n    "typeName": "Stack",\n    "props": {\n      "children": [\n        {\n          "type": "element",\n          "typeName": "Card",\n          "props": {\n            "children": [\n              {\n                "type": "element",\n                "typeName": "CardHeader",\n                "props": {\n                  "title": "👩‍💻  Maya Chen",\n                  "subtitle": "Senior Frontend Engineer"\n                },\n                "partial": false,\n                "hasDynamicProps": false\n              },\n              {\n                "type": "element",\n                "typeName": "Stack",\n                "props": {\n                  "children": [\n                    {\n                      "type": "element",\n                      "typeName": "Tag",\n                      "props": {\n                        "text": "📍  San Francisco"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Tag",\n                      "props": {\n                        "text": "⚛️  React"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Tag",\n                      "props": {\n                        "text": "🧩  Design Systems"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Tag",\n                      "props": {\n                        "text": "⏳  8 yrs exp"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    }\n                  ],\n                  "direction": "row",\n                  "wrap": true,\n                  "gap": "m",\n                  "align": "stretch",\n                  "justify": "start"\n                },\n                "partial": false,\n                "hasDynamicProps": false,\n                "statementId": "profileMeta"\n              }\n            ],\n            "variant": "card",\n            "direction": "column",\n            "wrap": false,\n            "gap": "m",\n            "align": "start",\n            "justify": "start"\n          },\n          "partial": false,\n          "hasDynamicProps": false,\n          "statementId": "profileCard"\n        },\n        {\n          "type": "element",\n          "typeName": "Card",\n          "props": {\n            "children": [\n              {\n                "type": "element",\n                "typeName": "CardHeader",\n                "props": {\n                  "title": "📊  Stats",\n                  "subtitle": "At a glance"\n                },\n                "partial": false,\n                "hasDynamicProps": false\n              },\n              {\n                "type": "element",\n                "typeName": "Stack",\n                "props": {\n                  "children": [\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "8.4k",\n                              "size": "large-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "👥  followers",\n                              "size": "small"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": 142,\n                              "size": "large-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "📝  posts",\n                              "size": "small"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": 312,\n                              "size": "large-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "➡️  following",\n                              "size": "small"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false\n                    }\n                  ],\n                  "direction": "row",\n                  "wrap": true,\n                  "gap": "l",\n                  "align": "stretch",\n                  "justify": "start"\n                },\n                "partial": false,\n                "hasDynamicProps": false,\n                "statementId": "statsRow"\n              }\n            ],\n            "variant": "sunk",\n            "direction": "column",\n            "wrap": false,\n            "gap": "m",\n            "align": "start",\n            "justify": "start"\n          },\n          "partial": false,\n          "hasDynamicProps": false,\n          "statementId": "statsCard"\n        },\n        {\n          "type": "element",\n          "typeName": "Card",\n          "props": {\n            "children": [\n              {\n                "type": "element",\n                "typeName": "CardHeader",\n                "props": {\n                  "title": "🪪  About",\n                  "subtitle": "A few words"\n                },\n                "partial": false,\n                "hasDynamicProps": false\n              },\n              {\n                "type": "element",\n                "typeName": "TextContent",\n                "props": {\n                  "text": "Building thoughtful interfaces with React and TypeScript. Previously at Linear and Stripe. Open-source contributor focused on design systems and developer tooling.",\n                  "size": "default"\n                },\n                "partial": false,\n                "hasDynamicProps": false,\n                "statementId": "bioText"\n              }\n            ],\n            "variant": "card",\n            "direction": "column",\n            "wrap": false,\n            "gap": "m",\n            "align": "start",\n            "justify": "start"\n          },\n          "partial": false,\n          "hasDynamicProps": false,\n          "statementId": "bioCard"\n        },\n        {\n          "type": "element",\n          "typeName": "Card",\n          "props": {\n            "children": [\n              {\n                "type": "element",\n                "typeName": "CardHeader",\n                "props": {\n                  "title": "🕒  Recent Activity",\n                  "subtitle": "What she\'s been up to"\n                },\n                "partial": false,\n                "hasDynamicProps": false\n              },\n              {\n                "type": "element",\n                "typeName": "Stack",\n                "props": {\n                  "children": [\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "💬  Replied to a thread on design systems",\n                              "size": "small-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "2 hours ago",\n                              "size": "small"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false,\n                      "statementId": "act1"\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Separator",\n                      "props": {},\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "🚀  Shipped v2.0 of use-open-ui",\n                              "size": "small-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "Yesterday",\n                              "size": "default"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false,\n                      "statementId": "act2"\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Separator",\n                      "props": {},\n                      "partial": false,\n                      "hasDynamicProps": false\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Stack",\n                      "props": {\n                        "children": [\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "✍️  Published \'Edge rendering with Lynx\'",\n                              "size": "small-heavy"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          },\n                          {\n                            "type": "element",\n                            "typeName": "TextContent",\n                            "props": {\n                              "text": "3 days ago",\n                              "size": "small"\n                            },\n                            "partial": false,\n                            "hasDynamicProps": false\n                          }\n                        ],\n                        "direction": "column",\n                        "wrap": false,\n                        "gap": "xs"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": false,\n                      "statementId": "act3"\n                    }\n                  ],\n                  "direction": "column",\n                  "wrap": false,\n                  "gap": "s"\n                },\n                "partial": false,\n                "hasDynamicProps": false,\n                "statementId": "activityList"\n              }\n            ],\n            "variant": "card",\n            "direction": "column",\n            "wrap": false,\n            "gap": "m",\n            "align": "stretch",\n            "justify": "start"\n          },\n          "partial": false,\n          "hasDynamicProps": false,\n          "statementId": "activityCard"\n        },\n        {\n          "type": "element",\n          "typeName": "Stack",\n          "props": {\n            "children": [\n              {\n                "type": "element",\n                "typeName": "Buttons",\n                "props": {\n                  "buttons": [\n                    {\n                      "type": "element",\n                      "typeName": "Button",\n                      "props": {\n                        "label": "➕  Follow",\n                        "action": {\n                          "k": "Comp",\n                          "name": "Action",\n                          "args": [\n                            {\n                              "k": "Arr",\n                              "els": [\n                                {\n                                  "k": "Comp",\n                                  "name": "ToAssistant",\n                                  "args": [\n                                    {\n                                      "k": "Str",\n                                      "v": "Follow Maya Chen"\n                                    }\n                                  ]\n                                }\n                              ]\n                            }\n                          ]\n                        },\n                        "variant": "primary",\n                        "type": "normal",\n                        "size": "large"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": true\n                    },\n                    {\n                      "type": "element",\n                      "typeName": "Button",\n                      "props": {\n                        "label": "✉️  Message",\n                        "action": {\n                          "k": "Comp",\n                          "name": "Action",\n                          "args": [\n                            {\n                              "k": "Arr",\n                              "els": [\n                                {\n                                  "k": "Comp",\n                                  "name": "ToAssistant",\n                                  "args": [\n                                    {\n                                      "k": "Str",\n                                      "v": "Send Maya a message"\n                                    }\n                                  ]\n                                }\n                              ]\n                            }\n                          ]\n                        },\n                        "variant": "secondary",\n                        "type": "normal",\n                        "size": "medium"\n                      },\n                      "partial": false,\n                      "hasDynamicProps": true\n                    }\n                  ]\n                },\n                "partial": false,\n                "hasDynamicProps": true\n              }\n            ],\n            "direction": "row",\n            "wrap": false,\n            "gap": "m",\n            "align": "start",\n            "justify": "start"\n          },\n          "partial": false,\n          "hasDynamicProps": true,\n          "statementId": "actionRow"\n        }\n      ],\n      "direction": "column",\n      "wrap": false,\n      "gap": "l"\n    },\n    "partial": false,\n    "hasDynamicProps": true,\n    "statementId": "root"\n  },\n  "meta": {\n    "incomplete": false,\n    "unresolved": [],\n    "orphaned": [],\n    "statementCount": 13,\n    "errors": []\n  },\n  "stateDeclarations": {},\n  "queryStatements": [],\n  "mutationStatements": []\n}',
   },
   {
     id: 'simple-hello-world',
