+++ packages/genui/playground/src/pages/chat/shared.ts
@@ -10,7 +10,6 @@ import type {
 import type { ProtocolName } from '../../utils/protocol.js';
 import { isDevHost } from '../../utils/publishPayload.js';
 
-export const ONLINE_GENUI_SERVER_ORIGIN = 'https://genui-server.vercel.app';
 export const LOCAL_GENUI_SERVER_PORT = '3060';
 
 export const CHAT_PROVIDER_SETTINGS_STORAGE_KEY =
@@ -303,9 +302,7 @@ export function resolveTrustedChatEndpoint(
   try {
     const endpoint = new URL(raw, host.origin);
     if (endpoint.origin === host.origin) return endpoint.toString();
-    if (endpoint.origin === ONLINE_GENUI_SERVER_ORIGIN) {
-      return endpoint.toString();
-    }
+    if (endpoint.protocol === 'https:') return endpoint.toString();
 
     const isTrustedDevEndpoint = endpoint.protocol === 'http:'
       && endpoint.port === LOCAL_GENUI_SERVER_PORT
@@ -330,7 +327,7 @@ export function getChatEndpoint(
   if (host.protocol === 'http:' && isDevHost(host.hostname)) {
     return `http://${host.hostname}:${LOCAL_GENUI_SERVER_PORT}/${protocol}/stream`;
   }
-  return `${ONLINE_GENUI_SERVER_ORIGIN}/${protocol}/stream`;
+  return `${host.origin}/${protocol}/stream`;
 }
 
 export function getA2UIActionEndpoint(chatEndpoint: string): string {
