+++ packages/genui/playground/src/utils/publishPayload.ts
@@ -2,7 +2,6 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-const ONLINE_A2UI_SERVER_ORIGIN = 'https://genui-server.vercel.app';
 const LOCAL_A2UI_SERVER_PORT = '3060';
 
 export interface PublishedPayload {
@@ -31,7 +30,7 @@ export function getA2UIPayloadEndpoint(): string {
   ) {
     return `http://${window.location.hostname}:${LOCAL_A2UI_SERVER_PORT}/a2ui/payload`;
   }
-  return `${ONLINE_A2UI_SERVER_ORIGIN}/a2ui/payload`;
+  return new URL('/a2ui/payload', window.location.origin).toString();
 }
 
 export function getOpenUIPayloadEndpoint(): string {
@@ -40,7 +39,7 @@ export function getOpenUIPayloadEndpoint(): string {
   ) {
     return `http://${window.location.hostname}:${LOCAL_A2UI_SERVER_PORT}/openui/payload`;
   }
-  return `${ONLINE_A2UI_SERVER_ORIGIN}/openui/payload`;
+  return new URL('/openui/payload', window.location.origin).toString();
 }
 
 /**
