+++ packages/genui/server/package.json
@@ -7,7 +7,7 @@
     "build": "tsc --noEmit && rslib build",
     "dev": "node scripts/dev.mjs",
     "lint": "eslint --ignore-pattern dist .",
-    "start": "node dist/index.js",
+    "start": "node dist/server.js",
     "test": "rstest"
   },
   "dependencies": {
